--- Packages/com.careboo.blinq/CareBoo.Blinq/OrderedValueSequence.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4c0a6433ad5d9843ad28b900bb1171d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
