--- Packages/com.careboo.blinq/CareBoo.Blinq/SameSelector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ec85e71345600b649a323b701d62dfc6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
